% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doocall.R
\name{doocall}
\alias{doocall}
\title{doocall}
\usage{
doocall(.what, ...)
}
\arguments{
\item{.what}{what parameter for do.call}

\item{...}{pass additional arguments to do.call}
}
\value{
results of the .what function
}
\description{
constructs and executes a function call from a name or a function (see do.call)
and automatically uses the arguments from the function provided in .what
}
\examples{
summer <- function(x,y){sum(x,y)}
x <- 1:5
y <- 1:5
z_doocall <- doocall(.what = summer)
}
